restore padding to base64-encoded service_name in UpstreamService
