+++ src/base64_rfc4648.h
@@ -9,10 +9,10 @@ namespace datadog {
 namespace opentracing {
 
 // Encode the specified `source` data as base64, where the encoding is
-// compatible with RFC 4648 _except_ that any trailing padding characters are
-// omitted.  Append to the specified `destination` the encoded result,
-// expanding the storage of `destination` as needed.
-void appendBase64Unpadded(std::string& destination, ot::string_view source);
+// compatible with RFC 4648, including padding.  Append to the specified
+// `destination` the encoded result, expanding the storage of `destination` as
+// needed.
+void appendBase64(std::string& destination, ot::string_view source);
 
 }  // namespace opentracing
 }  // namespace datadog
