+++ test/upstream_service_test.cpp
@@ -50,17 +50,17 @@ TEST_CASE("appendUpstreamService") {
     },
     { // just one service
       {{"mysvc", SamplingPriority::UserKeep, int(SamplingMechanism::Rule), 0.01234}},
-      "bXlzdmM|2|3|0.0124"
+      "bXlzdmM=|2|3|0.0124"
     },
     { // two services
       {{"yoursvc", SamplingPriority::SamplerDrop, 1337, 1.0},
        {"mysvc", SamplingPriority::UserKeep, int(SamplingMechanism::Rule), 0.01234}},
-      "eW91cnN2Yw|0|1337|1.0000;bXlzdmM|2|3|0.0124"
+      "eW91cnN2Yw==|0|1337|1.0000;bXlzdmM=|2|3|0.0124"
     },
     { // example based on internal design document
       {{"mcnulty-web", SamplingPriority::SamplerDrop, int(SamplingMechanism::AgentRate), std::nan("")},
        {"trace-stats-query", SamplingPriority::UserKeep, int(SamplingMechanism::Manual), std::nan("")}},
-      "bWNudWx0eS13ZWI|0|1|;dHJhY2Utc3RhdHMtcXVlcnk|2|4|"
+      "bWNudWx0eS13ZWI=|0|1|;dHJhY2Utc3RhdHMtcXVlcnk=|2|4|"
     }
   }));
   // clang-format on
